added a for loop
The for loop keeps the player from pressing higher numbers than there are states in the the array. I think it's interesting that this works, the game loop must cache the last button press because otherwise what are the odds the for loop is in the same iteration as the button that you pressed?